feat: Added user side translation # User can type and submit question and see the list of already answered faq's.
# User can change the language from the
select language dropdown and see the
answered faq's in any language it wants.

# The backend works in such a manner that it
first stores the typed question from the
frontend in the database and when the user
changes the language, the backend picks up
every question and answer from the database,
translates it using google translate api and
sends the translated text to frontend and
translated texts are displayed on the screen.